+++ src/gl-state-wgl.h
@@ -38,7 +38,7 @@ public:
     bool valid();
     bool reset();
     void swap();
-    bool gotNativeConfig(int& vid);
+    bool gotNativeConfig(intptr_t& vid);
     void getVisualConfig(GLVisualConfig& vc);
 
 private:
